refactor: simplify timeline components
