+++ src/main.jsx
@@ -2,7 +2,6 @@ import React, { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { useSelectionStore } from "@/shared/store"; // 초기화 용도
 import App from "@/App";
 import "@/index.css";
